implementation of modular loading functons done
