+++ deflib1.py
@@ -13,6 +13,8 @@ import matplotlib.pyplot as plt
 import matplotlib.patches as mpatches
 from matplotlib.patches import Rectangle
 import loadXYspecs as specloader
+import mathlib3 as matl
+import error_handler
 
 Notebooks = ['Load Data', 'Hyperspectra', 'Clara Image', 'HSI Plot', 'Newton Spectrum', 'TCSPC', 'HSI File Sorter', 'Plot Spectra', 'Cube2Image', 'Settings', 'Marwins Specplotter']
 Loadingmethods = specloader.loadingmethods
@@ -20,6 +22,20 @@ Loadingmethods = specloader.loadingmethods
 DIR = os.path.dirname(os.path.abspath(__file__))
 # Default values for the application
 DEFAULTS_FILE = f'{DIR}/defaults.txt'
+
+
+SpectDataFloats = ['Slit Width (µm)', 'Central Wavelength (nm)',
+                   'Cooling Temperature (°C)',
+                   'Exposure Time (s)',
+                   #'Wavelength First Pixel (nm)', # is obtained from WL
+                   #'Wavelength Last Pixel (nm)',
+                   'Delta Wavelength (nm)',
+                   'x-position',
+                   'y-position',
+                   'z-position',
+                   'Short Wavelength (nm)',
+                   'Long Wavelength (nm)',
+                   'magnification']
 
 # load defaults
 def initdefaults():
@@ -1302,27 +1318,25 @@ def fig_on_hoverevent(event, ax, fig, Z, x_range, y_range):
                 ax.set_title(f"x: {event.xdata:.2f}, y: {event.ydata:.2f}, z: {z_value:.2f}")
                 fig.canvas.draw_idle()
 
-# todo: add a method that uses linear interpolation first and afterwards applies Nearest Neighbor average for remaining cosmics
-
 cosmicfuncts = {
-                'Linear Interpolation': remove_cosmics_linear, 
-                'Median Filter': remove_cosmics_median_filter, 
-                #'Rolling Mean': remove_cosmics_rolling_mean, 
-                #'Spline Interpolation': remove_cosmics_spline,
-                'Nearest Neighbor average': remove_cosmics_nearest_neighbor,
-                'Linear + Neighbor (Combined)': remove_cosmics_linear_then_neighbor,
-                'Matrix Image correction': matrix_image_correction, 
-                'Cosmic Correlation': cosmic_correlation,
-                'Adaptive Threshold': adaptive_threshold,
-                'Spectral Correlation': spectral_correlation,
-                'Bilateral Filter': bilateral_filter,
-                'Robust Median': robust_median,
-                'Iterative Cosmic': iterative_cosmic,
-                'PCA Anomaly': pca_anomaly,
-                'Gradient Based': gradient_based, 
-                'spacial removal': remove_cosmics_spacial,
-                'withaker-hayes despike': whitaker_hayes_despike
-                } 
+    'Linear Interpolation': remove_cosmics_linear, 
+    'Median Filter': remove_cosmics_median_filter, 
+    #'Rolling Mean': remove_cosmics_rolling_mean, 
+    #'Spline Interpolation': remove_cosmics_spline,
+    'Nearest Neighbor average': remove_cosmics_nearest_neighbor,
+    'Linear + Neighbor (Combined)': remove_cosmics_linear_then_neighbor,
+    'Matrix Image correction': matrix_image_correction, 
+    'Cosmic Correlation': cosmic_correlation,
+    'Adaptive Threshold': adaptive_threshold,
+    'Spectral Correlation': spectral_correlation,
+    'Bilateral Filter': bilateral_filter,
+    'Robust Median': robust_median,
+    'Iterative Cosmic': iterative_cosmic,
+    'PCA Anomaly': pca_anomaly,
+    'Gradient Based': gradient_based, 
+    'spacial removal': remove_cosmics_spacial,
+    'withaker-hayes despike': whitaker_hayes_despike
+} 
 
 # Correlation functions for cosmic ray removal must be run by XYMap class instead of SpectrumData class only
 correlationcosmicfuncts = {
@@ -1611,6 +1625,199 @@ def returnallfolders(maindir):
             allfolders.append(os.path.join(root, dir))
     return allfolders
 
+class SpectrumData:
+    def __init__(self, filename, WL, BG, loadeachbg = False, linearbg=False, removecosmics=False, cosmicthreshold=20, cosmicpixels=3, removecosmicmethod=list(cosmicfuncts.keys())[0], WL_eV=None, default_dataset='Spectrum (PL-BG)'):
+        self.removecosmicsmethod = removecosmicmethod
+        self.loadeachbg = loadeachbg
+        self.linearbg = linearbg
+        self.removecosmics = removecosmics
+        self.cosmicthreshold = cosmicthreshold
+        self.cosmicpixels = cosmicpixels
+        self.WL = WL
+        self.WL_eV = WL_eV
+        self.dofit = False # set True if fit is done
+        if self.loadeachbg == True:
+            self.BG = []
+        else:
+            self.BG = BG
+        self.filename = filename
+        self.readinkeys = ['BG', 'PL'] # WL is defined in XYMap
+        self.default_dataset = default_dataset
+        self.openFstate = [] # open floats state from metadata
+        self.openDstate = [] # open Data from spectrometer
+        self.dataokay = False    # set True if everything openend properly
+        self.data = {}
+        self.roistore = {}
+        self.PL = []
+        self.PLB = [] # PL-BG
+        self.Specdiff1 = None # first derivative dPLB/dWL (None until calculated)
+        self.Specdiff2 = None # second derivative d2PLB/dWL2 (None until calculated)
+        self._read_file()
+
+        # init fit data
+        self.fwhm = np.nan
+        self.fitmaxX = np.nan
+        self.fitmaxY = np.nan
+        self.fitdata = [None] # only fit parameter are stored, not the fit itself
+        # what is inside fitdata after the fit is done? 
+        # a sub array for each fit function (look at matl.fitkeys)
+        # each array contains the fit output parameters (fitkeys[function][4], which is the number of parameters), followed by:
+        # ['ss_res', 'ss_tot', 'r_squared', 'fwhm', 'pixstart', 'pixend', 'wlstart', 'wlend', 'max_x', 'max_y']
+        #  fitmaxX[-1] fitmaxY[-2] aqpixstart[-3] aqpixend[-4] r_squared[-7] 'ss_res'[-8] 'ss_tot'[-9] 
+        self.fitparams = matl.buildfitparas() # store all fit parameters
+        self.fitparamunits = matl.buildfitparas()
+
+    def _read_file(self):
+        """
+        Read spectrum data from file.
+        
+        Uses centralized error handling via ErrorEngine when available.
+        Uses get_default_error_engine() to access shared error handler instance.
+        Maintains dataokay flag to signal file read success/failure.
+        """
+        # Get default error engine (shared singleton instance)
+        # This is safe to use from any thread/context
+        error_engine = error_handler.get_default_error_engine()
+        #default_decoding = 'utf-8'  # Default encoding for reading files
+
+        try:
+            with open(self.filename, 'r') as file:
+                lines = file.readlines()
+        except Exception as e:
+            # Use ErrorEngine for file open errors
+            error_engine.error(
+                exception=e,
+                context="Opening spectrum file",
+                filename=self.filename,
+                action="file_open"
+            )
+            # Maintain existing dataokay behavior
+            self.dataokay = False
+            return
+
+        # Process lines to store variables
+        startreaddata = False
+        for line in lines:
+            if ':' in line:
+                key, value = map(str.strip, line.split(':', 1))
+                if key in SpectDataFloats:
+                    try:
+                        self.data[key] = float(value)
+                    except Exception as e:
+                        # Use ErrorEngine for parsing errors
+                        error_engine.warning(
+                            message=f"Could not parse float value for key '{key}'",
+                            context="Parsing spectrum metadata",
+                            filename=self.filename,
+                            key=key,
+                            value=value
+                        )
+                        self.data[key] = value
+                        self.openFstate.append(False)
+                else:
+                    self.data[key] = value
+            elif '\t' in line:  # Data lines with tabs
+                parts = line.split()
+                if startreaddata == False:
+                    count = 0
+                    # start reading if at least two keys in the line
+                    for i in parts:
+                        if i in self.readinkeys:
+                            count += 1
+                    if count > 1:
+                        startreaddata = True
+                elif startreaddata == True:
+                    try:
+                        if self.loadeachbg == True:
+                            #self.BG.append(int(parts[1]))
+                            print("Warning: loadeachbg is True, but BG data is not being loaded. Check implementation.")
+                        #self.WL.append(float(parts[0]))  WL is only read once by XYMap since each SpectrumData has the same WL-axis
+                        self.PL.append(int(parts[2])) # type: ignore
+                    except Exception as e:
+                        # Use ErrorEngine for data parsing errors
+                        error_engine.error(
+                            exception=e,
+                            context="Parsing spectrum data line",
+                            filename=self.filename,
+                            line_content=line.strip()
+                        )
+        if self.loadeachbg == True and self.linearbg == True:
+            av = np.mean(self.BG)
+            for i in range(len(self.BG)):
+                self.BG[i] = av
+
+        # Convert PL (and BG for loadeachbg) to float32 numpy arrays to reduce RAM usage
+        self.PL = np.array(self.PL, dtype=np.float32)
+        if self.loadeachbg == True:
+            self.BG = np.array(self.BG, dtype=np.float32)
+
+        try:
+            self.PLB = np.subtract(self.PL, self.BG).astype(np.float32) # add PLB = PL-BG
+        except Exception as e:
+            # Use ErrorEngine for background subtraction errors
+            error_engine.error(
+                exception=e,
+                context="Background subtraction",
+                filename=self.filename,
+                PL_length=len(self.PL),
+                BG_length=len(self.BG)
+            )
+        # write openstate list
+        for i in SpectDataFloats:
+            if i not in list(self.data.keys()):
+                self.openFstate.append(False)
+        if len(self.WL) == 0:
+            self.openDstate.append(None)
+        if len(self.BG) == 0:
+            self.openDstate.append(None)
+        if len(self.PL) == 0:
+            self.openDstate.append(None)
+        self.setOK()
+        # remove cosmic rays todo: add nearest neighbor method
+        if self.removecosmics == True:
+            try:
+                self.PLB = np.asarray(cosmicfuncts[self.removecosmicsmethod](self.PLB, self.cosmicthreshold, self.cosmicpixels), dtype=np.float32)
+            except Exception as e:
+                # Use ErrorEngine for cosmic removal errors
+                error_engine.error(
+                    exception=e,
+                    context="Cosmic ray removal",
+                    filename=self.filename,
+                    method=self.removecosmicsmethod,
+                    threshold=self.cosmicthreshold
+                )
+        
+        # if 
+        
+        # important: clean up! delete everything that is not needed anymore
+        del lines
+
+    def setOK(self):
+        if False in self.openFstate:
+            pass
+        elif None in self.openDstate:
+            pass
+        else:
+            self.dataokay = True
+
+    # return a specific attribute of this class
+    def get_attribute(self, attr_name:str):
+        try:
+            return getattr(self, attr_name)
+        except AttributeError as e:
+            print("Attribute {} not found in class SpectrumData.".format(attr_name))
+    
+    def __del__(self):
+        """Destructor to clean up resources"""
+        # Clear large data arrays to help garbage collector
+        if hasattr(self, 'PL'):
+            self.PL = []
+        if hasattr(self, 'BG'):
+            self.BG = []
+        if hasattr(self, 'PLB'):
+            self.PLB = []
+        # Note: WL is shared reference, don't clear it
+
 class HSIPlotManager:
     """Manage creation and storage of HSI plots using the existing plot_HSI function.
 
@@ -1738,6 +1945,7 @@ class HSIPlotManager:
                     entry['fig'].show()
                 except Exception:
                     pass
+
 
 
 # Example usage (put this in a new cell or run interactively):
